src/requirements.txt
dash==2.14.1
Flask==2.3.2
Flask-Login==0.6.3
pandas==2.1.4
plotly==5.18.0
gunicorn==21.2.0
python-dotenv==1.0.0